Add KUpCCGSD ansatz
